% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{asset1}
\alias{asset1}
\title{Kobotoolbox asset used in the introduction vignette}
\format{
A kobo_asset
}
\usage{
asset1
}
\description{
Kobotoolbox asset used in the introduction vignette.
}
\details{
Kobotoolbox asset used in the introduction vignette
}
\keyword{datasets}
