% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{asset2}
\alias{asset2}
\title{Kobotoolbox asset used in the multi-lang vignette}
\format{
A kobo_asset
}
\usage{
asset2
}
\description{
Kobotoolbox asset used in the multi-lang vignette.
}
\details{
Kobotoolbox asset used in the multi-lang vignette
}
\keyword{datasets}
