% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{asset3}
\alias{asset3}
\title{Kobotoolbox asset used in the repeat group vignette}
\format{
A kobo_asset
}
\usage{
asset3
}
\description{
Kobotoolbox asset used in the repeat group vignette.
}
\details{
Kobotoolbox asset used in the repeat group vignette
}
\keyword{datasets}
