% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{asset_list1}
\alias{asset_list1}
\title{Kobotoolbox asset list for the introduction vignette}
\format{
A tibble with one row per Kobotoolbox asset
}
\usage{
asset_list1
}
\description{
Kobotoolbox asset list for the introduction vignette.
}
\details{
Kobotoolbox asset list for the introduction vignette
}
\keyword{datasets}
