% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data1}
\alias{data1}
\title{Kobotoolbox asset list for the introduction vignette}
\format{
A tibble with submissions of the asset1
}
\usage{
data1
}
\description{
Kobotoolbox asset list for the introduction vignette.
}
\details{
Kobotoolbox asset list for the introduction vignette
}
\keyword{datasets}
