% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data2_en}
\alias{data2_en}
\title{Kobotoolbox asset list for the multi-lang vignette}
\format{
A tibble with submissions of the asset2 in English
}
\usage{
data2_en
}
\description{
Kobotoolbox asset list for the multi-lang vignette.
}
\details{
Kobotoolbox asset list for the multi-lang vignette
}
\keyword{datasets}
