% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data3}
\alias{data3}
\title{Kobotoolbox asset list for the repeat group vignette}
\format{
A tibble with submissions of the asset3
}
\usage{
data3
}
\description{
Kobotoolbox asset list for the repeat group vignette.
}
\details{
Kobotoolbox asset list for the repeat group vignette
}
\keyword{datasets}
