% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{asset_list}
\alias{asset_list}
\alias{asset_ml}
\alias{asset_rg}
\alias{asset_spatial}
\alias{asset_sm_label}
\alias{asset_audit}
\title{Examples of KoboToolbox assets, list of assets}
\format{
asset_list: a \code{data.frame} of 28 rows and 7 columns with a list of projects

asset_ml: A \code{kobo_asset} object on a project using multiple languages.

asset_rg: A \code{kobo_asset} object on a project using repeat groups.

asset_spatial: A \code{kobo_asset} object on a project showcasing gps questions.

asset_spatial: A \code{kobo_asset} object to showcase select multiple label.

asset_audit: A \code{kobo_asset} object on a project with audit log enabled.
}
\usage{
asset_list

asset_ml

asset_rg

asset_spatial

asset_sm_label

asset_audit
}
\description{
Examples of KoboToolbox assets, list of assets.
}
\keyword{datasets}
