% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_ml_en}
\alias{data_ml_en}
\alias{data_ml_fr}
\alias{data_ml_ar}
\alias{data_ml_default}
\alias{data_rg}
\alias{data_spatial}
\alias{data_sm_label}
\alias{data_audit}
\title{Examples of KoboToolbox submissions data}
\format{
data_ml: A \code{data.frame} with submissions from \code{asset_ml} in English.

data_ml: A \code{data.frame} with submissions from \code{asset_ml} in French.

data_ml: A \code{data.frame} with submissions from \code{asset_ml} in Arabic

data_ml: A \code{data.frame} with submissions from \code{asset_ml} with the default language.

data_rg: A \code{dm} object with submissions from \code{asset_rg}

data_spatial: A \code{data.frame} with submissions from
the \code{asset_spatial} KoboToolbox API asset.

data_sm_label: A \code{data.frame} with submissions from
\code{asset_sm_label}.

data_audit: A \code{data.frame} with submissions from \code{asset_audit}.
}
\usage{
data_ml_en

data_ml_fr

data_ml_ar

data_ml_default

data_rg

data_spatial

data_sm_label

data_audit
}
\description{
Examples of KoboToolbox submissions data.
}
\keyword{datasets}
