% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kobo_asset.R
\name{kobo_asset_list}
\alias{kobo_asset_list}
\title{List all available KoboToolbox API assets}
\usage{
kobo_asset_list()
}
\value{
a list of kobo_asset
}
\description{
List all available KoboToolbox API assets
}
