% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kobo_asset.R
\name{kobo_asset_version}
\alias{kobo_asset_version}
\title{Get a specific KoboToolbox Asset version from an asset uid or \code{kobo_asset}}
\usage{
kobo_asset_version(x, version)
}
\arguments{
\item{x}{the unique identifier of a specific asset or an asset object}

\item{version}{version of the asset}
}
\value{
a kobo_asset object
}
\description{
Get a specific KoboToolbox Asset version from an asset uid or \code{kobo_asset}
}
\examples{
\dontrun{
kobo_setup()
asset_list <- kobo_asset_list()
uid <- asset_list$uid[1]
asset <- kobo_asset(uid)
asset_version_list <- kobo_asset_version_list(asset)
kobo_asset_version(asset, asset_version_list$uid[1])
}

}
