% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kobo_audit.R
\name{kobo_audit}
\alias{kobo_audit}
\title{Get all audit logs data from a KoboToolbox survey}
\usage{
kobo_audit(x)
}
\arguments{
\item{x}{the unique identifier of a specific asset (\code{character}) or
a \code{kobo_asset} object.}
}
\value{
A \code{data.frame}. It contains survey paradata from audit logs. It is useful
for data quality control, troubleshooting or workflow management.
}
\description{
Get all audit logs data from a KoboToolbox survey through a \code{kobo_asset} or
asset unique identifier.
}
\examples{
\dontrun{
kobo_setup()
uid <- "a9cwEQcbWqWzdA5eqkjRUWi"
asset <- kobo_asset(uid)
audit <- kobo_audit(asset)

if (require(dplyr)) {
 library(dplyr)
 glimpse(audit)
 }
}

}
