% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-cache.R
\name{kobo_cache_info}
\alias{kobo_cache_info}
\title{Get cache statistics}
\usage{
kobo_cache_info()
}
\value{
A list with cache statistics
}
\description{
Get cache statistics
}
\examples{
\dontrun{
kobo_cache_info()
}

}
