% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kobo_file.R
\name{kobo_file_download}
\alias{kobo_file_download}
\title{Download all uploaded files related to a KoboToolbox API asset}
\usage{
kobo_file_download(x, folder)
}
\arguments{
\item{x}{the asset uid or the \code{kobo_asset} object.}

\item{folder}{character, the folder where you store the downloaded files.
The working directory is the default folder.}
}
\value{
A vector of file paths.
}
\description{
Download all uploaded files related to a KoboToolbox API asset
}
\examples{
\dontrun{
kobo_setup()
uid <- "a9cwEQcbWqWzA5hzkjRUWi"
kobo_file_download(uid, folder = tempdir())
}

}
