% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kobo_form.R
\name{kobo_form.kobo_asset}
\alias{kobo_form.kobo_asset}
\title{Get the form used for this asset}
\usage{
\method{kobo_form}{kobo_asset}(asset)
}
\arguments{
\item{asset}{kobo_asset, the asset}
}
\value{
kobo_form, the project form
}
\description{
Get the form used for this asset
}
