% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kobo_setup.R
\name{kobo_settings}
\alias{kobo_settings}
\title{Get \code{robotoolbox} settings}
\usage{
kobo_settings()
}
\value{
a list, prints your server url and token
}
\description{
Print the current token and server url used.
}
\examples{
\dontrun{
 kobo_settings()
}

}
