% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kobo_submissions.R
\name{kobo_submissions.kobo_asset}
\alias{kobo_submissions.kobo_asset}
\title{Get all submissions from a project}
\usage{
\method{kobo_submissions}{kobo_asset}(asset)
}
\arguments{
\item{asset}{kobo_asset, the asset}
}
\value{
data.frame, all submissions
}
\description{
Get all submissions from a project
}
