% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kobo_token.R
\name{kobo_token}
\alias{kobo_token}
\title{Get your KoBoToolbox API token}
\usage{
kobo_token(username = NULL, password = NULL, url = NULL)
}
\arguments{
\item{username}{character, KoBoToolbox username}

\item{password}{character, KoBoToolbox password}

\item{url}{character, KoBoToolbox url}
}
\value{
character, the KoBoToolbox API token
}
\description{
Get your KoBoToolbox API token
}
