% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robotoolbox-package.R
\docType{package}
\name{robotoolbox-package}
\alias{robotoolbox}
\alias{robotoolbox-package}
\title{robotoolbox: Client for the 'KoboToolbox' API}
\description{
\code{\link{robotoolbox}} is a Client for the 'KoboToolbox' API.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://dickoa.gitlab.io/robotoolbox}
  \item \url{https://gitlab.com/dickoa/robotoolbox}
  \item Report bugs at \url{https://gitlab.com/dickoa/robotoolbox/-/issues}
}

}
\author{
\strong{Maintainer}: Ahmadou Dicko \email{mail@ahmadoudicko.com} [copyright holder]

Other contributors:
\itemize{
  \item Hisham Galal \email{galal@unhcr.org} [contributor]
}

}
\keyword{internal}
