% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_group.R
\name{var_group}
\alias{var_group}
\alias{var_group<-}
\title{Get / Set a variable group}
\usage{
var_group(x, unlist)

var_group(x) <- value
}
\arguments{
\item{x}{a vector or a data.frame}

\item{unlist}{for data frames, return a named vector instead of a list}

\item{value}{a character string or `NULL` to remove the group
For data frames, it could also be a named list or a character vector
of same length as the number of columns in `x`.}
}
\description{
Get / Set a variable group
}
\details{
For data frames, if `value` is a named list, only elements whose name will
  match a column of the data frame will be taken into account. If `value`
  is a character vector, groups should in the same order as the columns of the
  data.frame.
}
\examples{
var_group(iris$Sepal.Length)
var_group(iris$Sepal.Length) <- 'length group'
\dontrun{
 View(iris)
}
# To remove a variable group
var_group(iris$Sepal.Length) <- NULL
# To change several variable groups at once
var_group(iris) <- c(
  "length group", "width group", "length group",
  "width group", "species"
)
var_group(iris)
var_group(iris) <- list(
  Petal.Width = "width group",
  Petal.Length = "length group"
)
var_group(iris)
var_group(iris, unlist = TRUE)

}
