% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{asset_list}
\alias{asset_list}
\alias{asset_ml}
\alias{asset_rg}
\alias{asset_spatial}
\alias{asset_sm_label}
\alias{asset_audit}
\title{Examples of KoboToolbox assets and list of assets}
\format{
asset_list: a \code{data.frame} of 28 rows and 7 columns with a list of API assets

asset_ml: A \code{kobo_asset} object on a survey using multiple languages.

asset_rg: A \code{kobo_asset} object on a survey using repeat groups.

asset_spatial: A \code{kobo_asset} object on a survey showcasing gps questions.

asset_sm_label: A \code{kobo_asset} object to showcase select multiple labels.

asset_audit: A \code{kobo_asset} object on a survey with audit logging enabled.
}
\usage{
asset_list

asset_ml

asset_rg

asset_spatial

asset_sm_label

asset_audit
}
\description{
Examples of KoboToolbox assets and list of assets.
}
\keyword{datasets}
