% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{asset_list1}
\alias{asset_list1}
\alias{asset1}
\alias{asset2}
\alias{asset3}
\alias{data1}
\alias{data2_en}
\alias{data2_fr}
\alias{data2_ar}
\alias{data2_default}
\alias{data3}
\title{KoboToolbox asset list for the introduction vignette}
\format{
A data.frame with one row per KoboToolbox asset

A kobo_asset

A kobo_asset

A kobo_asset

A data.frame with submissions of the asset1

A data.frame with submissions of the asset2 in English

A data.frame with submissions of the asset2 in French

A data.frame with submissions of the asset2 in Arabic

A data.frame with submissions of the asset2

A data.frame with submissions of the asset3
}
\usage{
asset_list1

asset1

asset2

asset3

data1

data2_en

data2_fr

data2_ar

data2_default

data3
}
\description{
KoboToolbox asset list for the introduction vignette.

KoboToolbox asset used in the introduction vignette.

KoboToolbox asset used in the multi-lang vignette.

KoboToolbox asset used in the repeat group vignette.

KoboToolbox submissions for the introduction vignette.

KoboToolbox asset list for the multi-lang vignette.

KoboToolbox asset list for the multi-lang vignette.

KoboToolbox asset list for the multi-lang vignette.

KoboToolbox asset list for the multi-lang vignette.

KoboToolbox asset list for the repeat group vignette.
}
\details{
KoboToolbox asset list for the introduction vignette

KoboToolbox asset used in the introduction vignette

KoboToolbox asset used in the multi-lang vignette

KoboToolbox asset used in the repeat group vignette

KoboToolbox asset list for the introduction vignette

KoboToolbox asset list for the multi-lang vignette

KoboToolbox asset list for the multi-lang vignette

KoboToolbox asset list for the multi-lang vignette

KoboToolbox asset list for the multi-lang vignette

KoboToolbox asset list for the repeat group vignette
}
\keyword{datasets}
