% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{asset_list}
\alias{asset_list}
\alias{asset_ml}
\alias{asset_rg}
\alias{asset_spatial}
\alias{data_ml_en}
\alias{data_ml_fr}
\alias{data_ml_ar}
\alias{data_ml_default}
\alias{data_rg}
\alias{data_spatial}
\title{KoboToolbox asset, assets list and submissions data}
\format{
asset_list: a data.frame of 28 rows and 7 columns with a list of projects

asset_ml: A kobo_asset object on a project using multiple languages.

asset_rg: A kobo_asset object on a project using repeat groups.

asset_spatial: A kobo_asset object on a project showcasing gps questions.

data_ml: A data.frame with submissions from the \code{asset_ml} in English.

data_ml: A data.frame with submissions from the \code{asset_ml} in French.

data_ml: A data.frame with submissions from the \code{asset_ml} in Arabic

data_ml: A data.frame with submissions from the \code{asset_ml} with default language.

data_rg: A dm object (list of data.frame) with submissions from the \code{asset_rg}

data_spatial: A data.frame with submissions from the \code{asset_spatial}.
}
\usage{
asset_list

asset_ml

asset_rg

asset_spatial

data_ml_en

data_ml_fr

data_ml_ar

data_ml_default

data_rg

data_spatial
}
\description{
KoboToolbox asset, assets list and submissions data
}
\details{
Examples of KoboToolbox asset, assets list and submissions data
}
\keyword{datasets}
