% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kobo_asset.R
\name{kobo_asset}
\alias{kobo_asset}
\title{Get a specific KoBoToolbox Asset from a uid}
\usage{
kobo_asset(uid)
}
\arguments{
\item{uid}{the unique identifier of a specific asset}
}
\value{
a kobo_asset object
}
\description{
Get a specific KoBoToolbox Asset from a uid
}
