% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kobo_asset.R
\name{kobo_asset}
\alias{kobo_asset}
\title{Get a specific KoboToolbox Asset from a uid}
\usage{
kobo_asset(x)
}
\arguments{
\item{x}{the unique identifier of a specific asset or an asset object}
}
\value{
a \code{kobo_asset} object
}
\description{
Get a specific KoboToolbox Asset from a uid
}
\examples{
\dontrun{
kobo_setup() # setup using your url and token
uid <- "a9cwEQcbWqWzA5hzkjRUWi" # pick a valid uid
asset <- kobo_asset(uid)
asset
}
}
