% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kobo_asset.R
\name{kobo_asset}
\alias{kobo_asset}
\title{Get a specific KoboToolbox API asset from a unique identifier}
\usage{
kobo_asset(x)
}
\arguments{
\item{x}{the unique identifier of a specific asset (\code{character}) or
a \code{kobo_asset} object.}
}
\value{
A \code{kobo_asset} object. It contains all the information about the
KoboToolbox API asset associated to the unique identifier.
}
\description{
Get a specific KoboToolbox API asset from a unique identifier
}
\examples{
\dontrun{
# replace by your own url and token
kobo_setup(url = "https://kf.kobotoolbox.org", token = "abcde")
# use a valid uid
uid <- "a9cwEQcbWqWzA5hzkjRUWi"
asset <- kobo_asset(uid)
asset
}
}
