% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kobo_asset.R
\name{kobo_asset_list}
\alias{kobo_asset_list}
\title{List all available KoboToolbox API assets}
\usage{
kobo_asset_list(limit = 100L)
}
\arguments{
\item{limit}{integer, the number of project to display per page. Default to 100.}
}
\value{
A \code{data.frame} containing the list of all your KoboToolbox projects and some
associated metadata.
}
\description{
List all available KoboToolbox API assets and their metadata.
}
\details{
This function list all the assets (projects) in your Kobotoolbox account
with metadata such as \code{uid}, \code{name}, \code{asset_type}, \code{owner_username}, \code{date_created},
\code{date_modified} and \code{submissions}.
}
\examples{
\dontrun{
kobo_setup()
asset_list <- kobo_asset_list()
asset_list
}

}
