% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kobo_asset.R
\name{kobo_asset_list}
\alias{kobo_asset_list}
\title{List all available KoboToolbox API assets}
\usage{
kobo_asset_list(limit = 100L)
}
\arguments{
\item{limit}{integer, the number of project to display per page. Default to 100}
}
\value{
a list of \code{\link{kobo_asset}}
}
\description{
List all available KoboToolbox API assets
}
\examples{
\dontrun{
kobo_setup()
asset_list <- kobo_asset_list()
asset_list
}

}
