% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kobo_asset.R
\name{kobo_asset_version}
\alias{kobo_asset_version}
\alias{kobo_asset_version.character}
\alias{kobo_asset_version.kobo_asset}
\alias{kobo_asset_version.default}
\title{Get a specific KoBoToolbox Asset from a uid}
\usage{
kobo_asset_version(x, version)

\method{kobo_asset_version}{character}(x, version)

\method{kobo_asset_version}{kobo_asset}(x, version)

\method{kobo_asset_version}{default}(x, version)
}
\arguments{
\item{x}{the unique identifier of a specific asset or an asset object}

\item{version}{version of the asset}
}
\value{
a kobo_asset object
}
\description{
Get a specific KoBoToolbox Asset from a uid
}
