% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kobo_asset.R
\name{kobo_asset_version}
\alias{kobo_asset_version}
\title{Get a specific KoboToolbox Asset version from an asset uid or \code{\link{kobo_asset}}}
\usage{
kobo_asset_version(x, version)
}
\arguments{
\item{x}{the unique identifier of a specific asset or an asset object}

\item{version}{version of the asset}
}
\value{
a \code{kobo_asset_version} object
}
\description{
Get a specific KoboToolbox Asset version from an asset uid or \code{\link{kobo_asset}}
}
\examples{
\dontrun{
kobo_setup() # setup using your url and token
uid <- "a9cwEQcbWqWzA5hzkjRUWi" # pick a valid uid
asset <- kobo_asset(uid)
asset_version_list <- kobo_asset_version_list(asset)
kobo_asset_version(asset, asset_version_list$uid[1])
}

}
