% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kobo_asset.R
\name{kobo_asset_version_list}
\alias{kobo_asset_version_list}
\alias{kobo_asset_version_list.character}
\alias{kobo_asset_version_list.kobo_asset}
\alias{kobo_asset_version_list.default}
\title{List all available versions of a KoBoToolbox asset}
\usage{
kobo_asset_version_list(x)

\method{kobo_asset_version_list}{character}(x)

\method{kobo_asset_version_list}{kobo_asset}(x)

\method{kobo_asset_version_list}{default}(x)
}
\arguments{
\item{x}{the uid or kobo_asset object}
}
\value{
a list of kobo_asset_version
}
\description{
List all available versions of a KoBoToolbox asset. Works only if you own
the project.
}
