% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kobo_asset.R
\name{kobo_asset_version_list}
\alias{kobo_asset_version_list}
\title{List all available versions of a KoboToolbox asset}
\usage{
kobo_asset_version_list(x)
}
\arguments{
\item{x}{the uid or kobo_asset object}
}
\value{
a list of \code{\link{kobo_asset_version}}
}
\description{
List all available versions of a KoboToolbox asset. Works only if you own
the project.
}
\examples{
\dontrun{
kobo_setup() # setup using your url and token
uid <- "a9cwEQcbWqWzA5hzkjRUWi" # pick a valid uid
asset <- kobo_asset(uid)
kobo_asset_version_list(asset)
}

}
