% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kobo_audit.R
\name{kobo_audit}
\alias{kobo_audit}
\title{Get all audit log data from a KoboToolbox project}
\usage{
kobo_audit(x)
}
\arguments{
\item{x}{a \code{kobo_asset} or character, the asset}
}
\value{
A data.frame
}
\description{
Get all audit log data from a KoboToolbox project through a \code{\link{kobo_asset}} or asset uid.
}
\details{
\code{kobo_audit} allows you to access the audit log files and parse them
into a \code{data.frame} to make audit log data analysis easier.
}
\examples{
\dontrun{
kobo_setup() # setup using your url and token
uid <- "a9cwEQcbWqWzdA5eqkjRUWi" # pick a valid uid
asset <- kobo_asset(uid)
audit <- kobo_audit(asset)

if (require(dplyr)) {
 library(dplyr)
 glimpse(audit)
 }
}

}
