% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kobo_data.R
\name{kobo_data}
\alias{kobo_data}
\alias{kobo_submissions}
\alias{kobo_data.kobo_asset}
\alias{kobo_submissions.kobo_asset}
\alias{kobo_data.character}
\alias{kobo_submissions.character}
\alias{kobo_data.default}
\alias{kobo_submissions.default}
\title{Get all submissions from a project}
\usage{
kobo_data(x, paginate, page_size, lang)

kobo_submissions(x, paginate, page_size, lang)

\method{kobo_data}{kobo_asset}(x, paginate = FALSE, page_size = NULL, lang = NULL)

\method{kobo_submissions}{kobo_asset}(x, paginate = FALSE, page_size = NULL, lang = NULL)

\method{kobo_data}{character}(x, paginate = FALSE, page_size = NULL, lang = NULL)

\method{kobo_submissions}{character}(x, paginate = FALSE, page_size = NULL, lang = NULL)

\method{kobo_data}{default}(x, paginate = FALSE, page_size = NULL, lang = NULL)

\method{kobo_submissions}{default}(x, paginate = FALSE, page_size = NULL, lang = NULL)
}
\arguments{
\item{x}{a kobo_asset or asset uid, the asset}

\item{paginate}{logical, split submissions by page. Default to FALSE}

\item{page_size}{integer, number of submissions per page.
if missing, default to number of submissions divided by 5}

\item{lang}{character, language for the variable and value labels}
}
\value{
data.frame
}
\description{
Get all submissions from a project (KoboToolbox asset)
}
