% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kobo_data.R
\name{kobo_data}
\alias{kobo_data}
\alias{kobo_submissions}
\alias{kobo_submissions.kobo_asset}
\alias{kobo_submissions.character}
\alias{kobo_submissions.default}
\title{Get all submissions from a KoboToolbox project}
\usage{
kobo_data(x, lang, select_multiple_label, all_versions, paginate, page_size)

kobo_submissions(
  x,
  lang,
  select_multiple_label,
  all_versions,
  paginate,
  page_size
)

\method{kobo_submissions}{kobo_asset}(
  x,
  lang = NULL,
  select_multiple_label = FALSE,
  all_versions = TRUE,
  paginate = FALSE,
  page_size = NULL
)

\method{kobo_submissions}{character}(
  x,
  lang = NULL,
  select_multiple_label = FALSE,
  all_versions = TRUE,
  paginate = FALSE,
  page_size = NULL
)

\method{kobo_submissions}{default}(
  x,
  lang = NULL,
  select_multiple_label = FALSE,
  all_versions = TRUE,
  paginate = FALSE,
  page_size = NULL
)
}
\arguments{
\item{x}{a \code{\link{kobo_asset}} or character, the asset
if missing, default to number of submissions divided by 5}

\item{lang}{character, language for the variable and value labels}

\item{select_multiple_label}{logical, whether or not to replace select_multiple columns values by labels. Default to \code{FALSE}.}

\item{all_versions}{logical, whether or not to include data from all form versions.
Default to TRUE}

\item{paginate}{logical, split submissions by page_size. Default to FALSE}

\item{page_size}{integer, number of submissions per page.}
}
\value{
A \code{data.frame} or A \code{dm} object if you have repeat groups
}
\description{
Get all submissions from a KoboToolbox project through a \code{\link{kobo_asset}} or asset uid.
}
\examples{
\dontrun{
kobo_setup() # setup using your url and token
uid <- "a9cwEQcbWqWzA5hzkjRUWi" # pick a valid uid
asset <- kobo_asset(uid)
subs <- kobo_data(asset) ## kobo_submissions(asset)

if (require(dplyr)) {
 library(dplyr)
 glimpse(subs)
 }
}

}
