% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kobo_data.R
\name{kobo_data}
\alias{kobo_data}
\alias{kobo_submissions}
\alias{kobo_submissions.kobo_asset}
\alias{kobo_submissions.character}
\alias{kobo_submissions.default}
\title{Get all submissions from a KoboToolbox project}
\usage{
kobo_data(x, paginate, page_size, lang)

kobo_submissions(x, paginate, page_size, lang)

\method{kobo_submissions}{kobo_asset}(x, paginate = FALSE, page_size = NULL, lang = NULL)

\method{kobo_submissions}{character}(x, paginate = FALSE, page_size = NULL, lang = NULL)

\method{kobo_submissions}{default}(x, paginate = FALSE, page_size = NULL, lang = NULL)
}
\arguments{
\item{x}{a \code{\link{kobo_asset}} or character, the asset}

\item{paginate}{logical, split submissions by page. Default to FALSE}

\item{page_size}{integer, number of submissions per page.
if missing, default to number of submissions divided by 5}

\item{lang}{character, language for the variable and value labels}
}
\value{
A data.frame
}
\description{
Get all submissions from a KoboToolbox project through a \code{\link{kobo_asset}} or asset uid.
}
\examples{
\dontrun{
kobo_setup() # setup using your url and token
uid <- "a9cwEQcbWqWzA5hzkjRUWi" # pick a valid uid
asset <- kobo_asset(uid)
subs <- kobo_data(asset) ## kobo_submissions(asset)
library(dplyr)
glimpse(subs)
}

}
