% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kobo_form.R
\name{kobo_form}
\alias{kobo_form}
\title{Get a KoboToolbox project form}
\usage{
kobo_form(x, version)
}
\arguments{
\item{x}{the asset uid or the \code{kobo_asset} object}

\item{version}{character, uid of the version of the asset}
}
\value{
a \code{data.frame}
}
\description{
Get a KoboToolbox project form from a \code{\link{kobo_asset}} or asset uid.
}
\details{
\code{kobo_form} pull the form associated to specific \code{kobo_asset}. The output is table with information such as \code{name}, \code{label}, \code{lang}, and list-column representing form \code{choices}.
}
\examples{
\dontrun{
#' kobo_setup()
asset_list <- kobo_asset_list()
uid <- asset_list$uid[1]
asset <- kobo_asset(uid)
form <- kobo_form(uid)
}

}
