% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kobo_form.R
\name{kobo_form}
\alias{kobo_form}
\title{Get the form used for this asset}
\usage{
kobo_form(x, version)
}
\arguments{
\item{x}{the asset uid or the kobo_asset object}

\item{version}{character, uid of the version of the asset}
}
\value{
tbl_form, the project form
}
\description{
Get the form used for this asset
}
\examples{
\dontrun{
#' kobo_setup()
asset_list <- kobo_asset_list()
uid <- asset_list$uid[1]
asset <- kobo_asset(uid)
form <- kobo_form(uid)
# form <- kobo_form(asset)
}

}
