% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kobo_form.R
\name{kobo_form}
\alias{kobo_form}
\title{Get a KoboToolbox project form}
\usage{
kobo_form(x, version)
}
\arguments{
\item{x}{the asset uid or the kobo_asset object}

\item{version}{character, uid of the version of the asset}
}
\value{
a data.frame with information on name, label, lang, and list-column representing form choices
}
\description{
Get a KoboToolbox project form from a \code{\link{kobo_asset}} or asset uid.
}
\examples{
\dontrun{
#' kobo_setup()
asset_list <- kobo_asset_list()
uid <- asset_list$uid[1]
asset <- kobo_asset(uid)
form <- kobo_form(uid)
# form <- kobo_form(asset)
}

}
