% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kobo_lang.R
\name{kobo_lang}
\alias{kobo_lang}
\alias{kobo_lang.kobo_asset}
\title{Languages used in the KoboToolbox project}
\usage{
kobo_lang(x)

\method{kobo_lang}{kobo_asset}(x)
}
\arguments{
\item{x}{kobo_asset or asset uid}
}
\value{
a vector of languages. It returnes "Labels" when no language is set

a vector of languages. It returnes "Labels" when no language is set
}
\description{
Languages used in the KoboToolbox project

Languages used in the KoboToolbox project
}
\examples{
\dontrun{
#' kobo_setup()
asset_list <- kobo_asset_list()
uid <- asset_list$uid[1]
asset <- kobo_asset(uid)
form <- kobo_form(uid)
# form <- kobo_form(asset)
}

\dontrun{
#' kobo_setup()
asset_list <- kobo_asset_list()
uid <- asset_list$uid[1]
asset <- kobo_asset(uid)
form <- kobo_form(uid)
# form <- kobo_form(asset)
}

}
