% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kobo_lang.R
\name{kobo_lang}
\alias{kobo_lang}
\alias{kobo_lang.kobo_asset}
\alias{kobo_lang.character}
\alias{kobo_lang.default}
\title{Languages used in the KoBoToolbox project}
\usage{
kobo_lang(x)

\method{kobo_lang}{kobo_asset}(x)

\method{kobo_lang}{character}(x)

\method{kobo_lang}{default}(x)
}
\arguments{
\item{x}{kobo_asset or asset uid}
}
\value{
a vector of languages. It returnes "Labels" when no language is set
}
\description{
Languages used in the KoBoToolbox project
}
