% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kobo_lang.R
\name{kobo_lang}
\alias{kobo_lang}
\title{Languages used in the KoboToolbox project}
\usage{
kobo_lang(x)
}
\arguments{
\item{x}{kobo_asset or asset uid, the asset}
}
\value{
a vector of languages used in the form. It returns "Labels" when no language is set
}
\description{
Languages used in the KoboToolbox project from a \code{\link{kobo_asset}} or asset uid.
}
\examples{
\dontrun{
kobo_setup() # setup using your url and token
uid <- "a9cwEQcbWqWzA5hzkjRUWi" # pick a valid uid
asset <- kobo_asset(uid)
lang <- kobo_lang(asset) ## or kobo_lang(uid)
lang
}

}
