% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kobo_lang.R
\name{kobo_lang}
\alias{kobo_lang}
\title{Get the languages used in a KoboToolbox survey form}
\usage{
kobo_lang(x)
}
\arguments{
\item{x}{the unique identifier of a specific asset (\code{character}) or
a \code{kobo_asset} object.}
}
\value{
A vector of \code{character}. The languages used in the form, it returns "Labels"
when no language is set.
}
\description{
Get the languages used in a KoboToolbox survey form from a \code{kobo_asset}
or asset unique identifier.
}
\examples{
\dontrun{
kobo_setup()
uid <- "a9cwEQcbWqWzA5hzkjRUWi"
asset <- kobo_asset(uid)
lang <- kobo_lang(asset)
lang
}

}
