% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kobo_setup.R
\name{kobo_settings}
\alias{kobo_settings}
\title{Get robotoolbox settings}
\usage{
kobo_settings()
}
\value{
A list with information about your KoboToolbox server URL and token.
}
\description{
Print the KoboToolbox server URL and API token currently in use.
}
\examples{
\dontrun{
 kobo_settings()
}

}
