% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kobo_setup.R
\name{kobo_setup}
\alias{kobo_setup}
\title{Set \code{robotoolbox} settings}
\usage{
kobo_setup(url = NULL, token = NULL)
}
\arguments{
\item{url}{the base url of the KoboToolbox server}

\item{token}{the API token}
}
\description{
Set \code{robotoolbox} settings
}
