% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kobo_setup.R
\name{kobo_setup}
\alias{kobo_setup}
\title{Set robotoolbox settings}
\usage{
kobo_setup(
  url = Sys.getenv("KOBOTOOLBOX_URL", ""),
  token = Sys.getenv("KOBOTOOLBOX_TOKEN", "")
)
}
\arguments{
\item{url}{character, the base URL of the KoboToolbox server.}

\item{token}{character, the API token.}
}
\value{
A \code{kobo_settings} object pritting the server URL and the API token.
}
\description{
Set the KoboToolbox server URL, API token and return invisibly
a \code{kobo_settings} object.
}
\examples{
\dontrun{
 # use your own URL and token
 kobo_setup(url = "https://kf.kobotoolbox.org/",
            token = "9et1814c285w094f6v9bd629df47a1a0e81x53a0")
 kobo_settings()
}

}
