% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kobo_submissions.R
\name{kobo_submissions}
\alias{kobo_submissions}
\alias{kobo_submissions.kobo_asset}
\title{Get all submissions from a project}
\usage{
kobo_submissions(asset, paginate, page_size)

\method{kobo_submissions}{kobo_asset}(asset, paginate = FALSE, page_size = NULL)
}
\arguments{
\item{asset}{kobo_asset, the asset}

\item{paginate}{logical, split submissions by page. Default to FALSE}

\item{page_size}{integer, number of submissions per page. if missing, default to
number of submissions divided by 5}
}
\value{
data.frame
}
\description{
Get all submissions from a project (KoBoToolbox asset)
}
