% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kobo_token.R
\name{kobo_token}
\alias{kobo_token}
\title{Get your KoBoToolbox API token from your username and password}
\usage{
kobo_token(username = NULL, password = NULL, url = NULL, overwrite = FALSE)
}
\arguments{
\item{username}{character, KoBoToolbox username}

\item{password}{character, KoBoToolbox password}

\item{url}{character, KoBoToolbox server url}

\item{overwrite}{logical, if TRUE overwrite existing token}
}
\value{
character, the KoBoToolbox API token
}
\description{
Get your KoBoToolbox API token from your username and password
}
