% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kobo_token.R
\name{kobo_token}
\alias{kobo_token}
\title{Get your KoboToolbox API token}
\usage{
kobo_token(username = NULL, password = NULL, url = NULL, overwrite = FALSE)
}
\arguments{
\item{username}{character, KoboToolbox username}

\item{password}{character, KoboToolbox password}

\item{url}{character, KoboToolbox server url}

\item{overwrite}{logical, if TRUE overwrite existing token}
}
\value{
character, the KoboToolbox API token
}
\description{
Get your KoboToolbox API token from your username and password.
}
\examples{
\dontrun{
token <- kobo_setup(username = "cool_user_name",
                    password = "xsww3@dddb",
                    url = "https://kf.kobotoolbox.org/")
token
}

}
