% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robotoolbox-package.R
\docType{package}
\name{robotoolbox}
\alias{robotoolbox}
\alias{robotoolbox-package}
\title{robotoolbox}
\description{
R Client for KoboToolbox API v2
}
\author{
\email{mail@ahmadoudicko.com}
}
\keyword{package}
