% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{set_names_from_varlabel}
\alias{set_names_from_varlabel}
\alias{set_names_from_varlabel.default}
\alias{set_names_from_varlabel.data.frame}
\alias{set_names_from_varlabel.dm}
\title{Use variable labels as variable names}
\usage{
set_names_from_varlabel(x)

\method{set_names_from_varlabel}{default}(x)

\method{set_names_from_varlabel}{data.frame}(x)

\method{set_names_from_varlabel}{dm}(x)
}
\arguments{
\item{x}{a tibble or dm object imported using \code{\link{kobo_data}}}
}
\description{
This function can be used alongside labelled::to_factor or labelled::to_character
to match the output you get from Kobotoolbox XLSX export with labels.
}
\examples{
\dontrun{
kobo_setup() # setup using your url and token
uid <- "a9cwEQcbWqWzA5hzkjRUWi" # pick a valid uid
asset <- kobo_asset(uid)
subs <- kobo_data(asset) ## kobo_submissions(asset)
if (require(dplyr)) {
 library(dplyr)
 glimpse(subs)
 glimpse(set_names_from_varlabel(subs))
 }
}

}
